+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2026-03-11
+
+- Added `maximumTakeoffMassKg` and `operatingFuelMassKg` properties to aircraft
+- Removed weight properties from aircraft which have none defined
+
 ## [1.5.1] - 2026-03-10
 
 - Added ICAO code to liveries which belong to known airlines